Refresh equation on submit
